pyproject.toml
[tool.poetry]
name = "loan-api"
version = "0.1.0"
description = ""
authors = ["MatheusLeall <matheuscardoso.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
drf-yasg = "^1.21.7"
coreapi = "^2.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
